Extract wheel-zoom listener into useGlobeWheel hook
